# src/NPaperless.Services/wwwroot/openapi-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Paperless Rest Server",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/api/document_types/" : {
      "get" : {
        "operationId" : "GetDocumentTypes",
        "parameters" : [ {
          "explode" : true,
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "full_perms",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "default" : {
                    "value" : "{\"count\":2,\"next\":null,\"previous\":null,\"all\":[1,2],\"results\":[{\"id\":1,\"slug\":\"testdoc\",\"name\":\"testdoc\",\"match\":\"\",\"matching_algorithm\":6,\"is_insensitive\":true,\"document_count\":1,\"owner\":3,\"permissions\":{\"view\":{\"users\":[3],\"groups\":[]},\"change\":{\"users\":[3],\"groups\":[]}}},{\"id\":2,\"slug\":\"testdocdel\",\"name\":\"testdocdel\",\"match\":\"del\",\"matching_algorithm\":1,\"is_insensitive\":true,\"document_count\":0,\"owner\":3,\"permissions\":{\"view\":{\"users\":[],\"groups\":[]},\"change\":{\"users\":[],\"groups\":[]}}}]}"
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/GetDocumentTypes_200_response"
                }
              }
            },
            "description" : "Success"
          }
        },
        "tags" : [ "DocumentTypes" ]
      },
      "post" : {
        "operationId" : "CreateDocumentType",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "examples" : {
                "default" : {
                  "value" : "{\"name\":\"testdocdel\",\"matching_algorithm\":1,\"match\":\"del\",\"is_insensitive\":true,\"owner\":3}"
                }
              },
              "schema" : {
                "$ref" : "#/components/schemas/CreateCorrespondent_request"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "default" : {
                    "value" : "{\"id\":2,\"slug\":\"testdocdel\",\"name\":\"testdocdel\",\"match\":\"del\",\"matching_algorithm\":1,\"is_insensitive\":true,\"owner\":3,\"user_can_change\":true}"
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/CreateDocumentType_200_response"
                }
              }
            },
            "description" : "Success"
          }
        },
        "tags" : [ "DocumentTypes" ]
      }
    },
    "/api/document_types/{id}/" : {
      "delete" : {
        "operationId" : "DeleteDocumentType",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "204" : {
            "description" : "Success"
          }
        },
        "tags" : [ "DocumentTypes" ]
      },
      "put" : {
        "operationId" : "UpdateDocumentType",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "examples" : {
                "default" : {
                  "value" : "{\"id\":2,\"slug\":\"testdocdel\",\"name\":\"testdocdel\",\"match\":\"del.*\",\"matching_algorithm\":4,\"is_insensitive\":true,\"document_count\":0,\"owner\":3,\"permissions\":{\"view\":{\"users\":[],\"groups\":[]},\"change\":{\"users\":[],\"groups\":[]}},\"set_permissions\":{\"view\":{\"users\":[],\"groups\":[]},\"change\":{\"users\":[],\"groups\":[]}},\"permissions_form\":{\"owner\":3,\"set_permissions\":{\"view\":{\"users\":[],\"groups\":[]},\"change\":{\"users\":[],\"groups\":[]}}}}"
                }
              },
              "schema" : {
                "$ref" : "#/components/schemas/UpdateDocumentType_request"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "default" : {
                    "value" : "{\"id\":2,\"slug\":\"testdocdel\",\"name\":\"testdocdel\",\"match\":\"del.*\",\"matching_algorithm\":4,\"is_insensitive\":true,\"document_count\":0,\"owner\":3,\"user_can_change\":true}"
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/UpdateDocumentType_200_response"
                }
              }
            },
            "description" : "Success"
          }
        },
        "tags" : [ "DocumentTypes" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "GetCorrespondents_200_response_results_inner_permissions_view" : {
        "example" : {
          "groups" : [ "", "" ],
          "users" : [ "", "" ]
        },
        "properties" : {
          "users" : {
            "items" : { },
            "type" : "array"
          },
          "groups" : {
            "items" : { },
            "type" : "array"
          }
        },
        "required" : [ "groups", "users" ],
        "type" : "object"
      },
      "GetCorrespondents_200_response_results_inner_permissions" : {
        "example" : {
          "view" : {
            "groups" : [ "", "" ],
            "users" : [ "", "" ]
          },
          "change" : {
            "groups" : [ "", "" ],
            "users" : [ "", "" ]
          }
        },
        "properties" : {
          "view" : {
            "$ref" : "#/components/schemas/GetCorrespondents_200_response_results_inner_permissions_view"
          },
          "change" : {
            "$ref" : "#/components/schemas/GetCorrespondents_200_response_results_inner_permissions_view"
          }
        },
        "required" : [ "change", "view" ],
        "type" : "object"
      },
...
      "CreateCorrespondent_request" : {
        "example" : {
          "owner" : 6,
          "matching_algorithm" : 0,
          "is_insensitive" : true,
          "name" : "name",
          "match" : "match"
        },
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "matching_algorithm" : {
            "type" : "integer"
          },
          "match" : {
            "type" : "string"
          },
          "is_insensitive" : {
            "type" : "boolean"
          },
          "owner" : {
            "type" : "integer"
          }
        },
        "required" : [ "is_insensitive", "match", "matching_algorithm", "name", "owner" ],
        "type" : "object"
      },
      "UpdateCorrespondent_request_permissions_form" : {
        "properties" : {
          "owner" : {
            "type" : "integer"
          },
          "set_permissions" : {
            "$ref" : "#/components/schemas/GetCorrespondents_200_response_results_inner_permissions"
          }
        },
        "required" : [ "owner", "set_permissions" ],
        "type" : "object"
      },
...
      "GetDocumentTypes_200_response_results_inner" : {
        "example" : {
          "owner" : 9,
          "matching_algorithm" : 2,
          "document_count" : 7,
          "is_insensitive" : true,
          "permissions" : {
            "view" : {
              "groups" : [ "", "" ],
              "users" : [ "", "" ]
            },
            "change" : {
              "groups" : [ "", "" ],
              "users" : [ "", "" ]
            }
          },
          "name" : "name",
          "match" : "match",
          "id" : 5,
          "slug" : "slug"
        },
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "slug" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "match" : {
            "type" : "string"
          },
          "matching_algorithm" : {
            "type" : "integer"
          },
          "is_insensitive" : {
            "type" : "boolean"
          },
          "document_count" : {
            "type" : "integer"
          },
          "owner" : {
            "type" : "integer"
          },
          "permissions" : {
            "$ref" : "#/components/schemas/GetCorrespondents_200_response_results_inner_permissions"
          }
        },
        "required" : [ "document_count", "id", "is_insensitive", "match", "matching_algorithm", "name", "owner", "permissions", "slug" ],
        "type" : "object"
      },
      "GetDocumentTypes_200_response" : {
        "example" : {
          "next" : 6,
          "all" : [ 5, 5 ],
          "previous" : 1,
          "count" : 0,
          "results" : [ {
            "owner" : 9,
            "matching_algorithm" : 2,
            "document_count" : 7,
            "is_insensitive" : true,
            "permissions" : {
              "view" : {
                "groups" : [ "", "" ],
                "users" : [ "", "" ]
              },
              "change" : {
                "groups" : [ "", "" ],
                "users" : [ "", "" ]
              }
            },
            "name" : "name",
            "match" : "match",
            "id" : 5,
            "slug" : "slug"
          }, {
            "owner" : 9,
            "matching_algorithm" : 2,
            "document_count" : 7,
            "is_insensitive" : true,
            "permissions" : {
              "view" : {
                "groups" : [ "", "" ],
                "users" : [ "", "" ]
              },
              "change" : {
                "groups" : [ "", "" ],
                "users" : [ "", "" ]
              }
            },
            "name" : "name",
            "match" : "match",
            "id" : 5,
            "slug" : "slug"
          } ]
        },
        "properties" : {
          "count" : {
            "type" : "integer"
          },
          "next" : {
            "type" : "integer"
          },
          "previous" : {
            "type" : "integer"
          },
          "all" : {
            "items" : {
              "type" : "integer"
            },
            "type" : "array"
          },
          "results" : {
            "items" : {
              "$ref" : "#/components/schemas/GetDocumentTypes_200_response_results_inner"
            },
            "type" : "array"
          }
        },
        "required" : [ "all", "count", "next", "previous", "results" ],
        "type" : "object"
      },
      "CreateDocumentType_200_response" : {
        "example" : {
          "owner" : 1,
          "matching_algorithm" : 6,
          "user_can_change" : true,
          "is_insensitive" : true,
          "name" : "name",
          "match" : "match",
          "id" : 0,
          "slug" : "slug"
        },
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "slug" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "match" : {
            "type" : "string"
          },
          "matching_algorithm" : {
            "type" : "integer"
          },
          "is_insensitive" : {
            "type" : "boolean"
          },
          "owner" : {
            "type" : "integer"
          },
          "user_can_change" : {
            "type" : "boolean"
          }
        },
        "required" : [ "id", "is_insensitive", "match", "matching_algorithm", "name", "owner", "slug", "user_can_change" ],
        "type" : "object"
      },
      "UpdateDocumentType_request" : {
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "slug" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "match" : {
            "type" : "string"
          },
          "matching_algorithm" : {
            "type" : "integer"
          },
          "is_insensitive" : {
            "type" : "boolean"
          },
          "document_count" : {
            "type" : "integer"
          },
          "owner" : {
            "type" : "integer"
          },
          "permissions" : {
            "$ref" : "#/components/schemas/GetCorrespondents_200_response_results_inner_permissions"
          },
          "set_permissions" : {
            "$ref" : "#/components/schemas/GetCorrespondents_200_response_results_inner_permissions"
          },
          "permissions_form" : {
            "$ref" : "#/components/schemas/UpdateCorrespondent_request_permissions_form"
          }
        },
        "required" : [ "document_count", "id", "is_insensitive", "match", "matching_algorithm", "name", "owner", "permissions", "permissions_form", "set_permissions", "slug" ],
        "type" : "object"
      },
      "UpdateDocumentType_200_response" : {
        "example" : {
          "owner" : 5,
          "matching_algorithm" : 6,
          "user_can_change" : true,
          "document_count" : 1,
          "is_insensitive" : true,
          "name" : "name",
          "match" : "match",
          "id" : 0,
          "slug" : "slug"
        },
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "slug" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "match" : {
            "type" : "string"
          },
          "matching_algorithm" : {
            "type" : "integer"
          },
          "is_insensitive" : {
            "type" : "boolean"
          },
          "document_count" : {
            "type" : "integer"
          },
          "owner" : {
            "type" : "integer"
          },
          "user_can_change" : {
            "type" : "boolean"
          }
        },
        "required" : [ "document_count", "id", "is_insensitive", "match", "matching_algorithm", "name", "owner", "slug", "user_can_change" ],
        "type" : "object"
      },
...
    }
  }
}